Moved sending message to popup logic into module
