+++ src/bandcamp/content.js
@@ -2,30 +2,6 @@
 
 let TralbumData, BandData = {};
 
-chrome.runtime.onMessage.addListener((request, sender, sendResponse) => {
-  let response;
-
-  if (request.type === 'getBandcampData') {
-    if (isOnReleasesListPage()) {
-      sendResponse(generateListResponse());
-    } else {
-      const currentTabUrl = window.location.href;
-      chrome.storage.local.get([currentTabUrl], (result) => {
-        if (result[currentTabUrl] && result[currentTabUrl]['release']) {
-          response = {
-            type: 'release',
-            data: result[currentTabUrl]['release']
-          };
-        }
-
-        sendResponse(response);
-      });
-    }
-  }
-
-  return true;
-});
-
 function extractReleaseData() {
   return {
     tralbumData: TralbumData,
@@ -42,44 +18,6 @@ function getSchemaData() {
   const scriptElement = document.querySelector('script[type="application/ld+json"]');
   const scriptContent = scriptElement.textContent;
   return JSON.parse(scriptContent);
-}
-
-function generateListResponse() {
-  const releases = [];
-  const releaseElements = document.querySelectorAll('#music-grid .music-grid-item');
-  const imgBandPhoto = document.querySelector('.band-photo');
-  const artistFilter = document.querySelector('#b2dArtistFilter');
-
-  releaseElements.forEach(element => {
-    let artist = element.querySelector('.artist-override')?.innerText;
-
-    if (!artist) {
-      artist = document.querySelector('#band-name-location .title').innerText;
-    }
-
-    const titleParts = element.querySelector('.title').innerText.split("\n");
-    const title = titleParts[0];
-    const url = element.querySelector('a').getAttribute('href');
-    releases.push({
-      url: (url[0] === '/' ? window.location.origin : '') + url,
-      artist: artist,
-      title: title
-    });
-  });
-
-  return {
-    type: 'list',
-    data: releases,
-    popup: {
-      imageSrc: imgBandPhoto.src,
-      search: artistFilter.value,
-    }
-  };
-}
-
-function isOnReleasesListPage() {
-  const location = window.location;
-  return location.pathname === '/' || location.pathname === '/music';
 }
 
 importModule("src/bandcamp/content_module.js");
